BUG/COMPAT: Fix compatibility and bugs associated with pandas toposort reimplementation
Author: Phillip Cloud <cpcloud@gmail.com>

Closes #1789 from cpcloud/add-back-main-execute-dispatch and squashes the following commits:

c8980d6 [Phillip Cloud] BUG/COMPAT: Fix compatibility and bugs associated with pandas toposort reimplementation

+++ ibis/pandas/execution/tests/test_window.py
@@ -269,7 +269,7 @@ def test_batting_cumulative_partitioned(batting, batting_df, sort_kind):
 
     tm.assert_series_equal(
         result.set_index([group_by, order_by]).sort_index().cumulative,
-        expected.sort_index(),
+        expected.sort_index().astype("int64"),
     )
